requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
dj-database-url==0.4.1
Django==2.1.0
django-cors-headers==2.0.0
django-cors-middleware==1.3.1
django-filter==2.0.0
django-rest-framework-social-oauth2==1.1.0
django-websocket-redis==0.4.7
djangorestframework==3.7.7
flake8==3.3.0
#gevent==1.1.2
greenlet==0.4.11
gunicorn==19.6.0
oauthlib==2.0.3
packaging==16.8
psycopg2==2.7.3.2
pyparsing==2.1.10
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
six==1.10.0
whitenoise==3.2.2
